move views related to auth to new directory
